requirements.txt
arrow==0.13.0
backcall==0.1.0
decorator==4.3.0
flake8==3.6.0
ipdb==0.11
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.2
kafka-python==1.4.4
mccabe==0.6.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycodestyle==2.4.0
pyflakes==2.0.0
Pygments==2.3.1
python-dateutil==2.7.5
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
